Modularize tasks code into versions and cache file to split up the big file a bit
