+++ dev-requirements.txt
@@ -1,2 +1,3 @@
 invoke>=2.2.0
 requests>=2.31.0
+cachetools>=5.3.0
